[DIRTY] Makefile: Output fs.img to ramdisk.h
This is FAR FAR from ideal. I mean truly.

What would be the ideal situation is to have
eMMC driver for the 2GB flash storage on
DshanNezha STU.

As you can imagine this is not happening.
That is due to few reasons. First and most
obvous. Im not porting an entire MMC driver
from linux kernel alongside initializing all
the clocks.

So what we have here is an array of all the
bytes in fs.img in header file which we can
later include and use to host all our files.

This does mean we do not have permanent storage
(unless... you want to leave the RAM ALWAYS POWERED
:P) and data will be removed upon power loss.

Signed-off-by: Martin Botka <martin.botka@somainline.org>