kernel: Remove references to virtio and add write flag to ramdisk
Removing virtio is obvious.
Adding a write flag isnt tho.

This flag tells the ramdisk read write function
if we should write or read. This is because the
B_VALID and B_DIRTY actually do not exist.

Also remove IRQ for virtio from trap.c
as we well do not need it.

It may not be pretty but oh well.

Signed-off-by: Martin Botka <martin.botka@somainline.org>