kernel: Switch CLINT to time CSR and initialize PMP
We have to initialize the PMP (Physical Memory Protection)
to actually be able to switch from M to S mode.

Also the CLINT register is changed on D1 and we have to
read the timestamp from time CSR

Write 1 to PLIC_CTRL address to be able to change PLIC config
in M mode. (Platform Level Interupt Controller)

Signed-off-by: Martin Botka <martin.botka@somainline.org>